[API] Refactor index.js into route files
Each API endpoint route now has it's own source file, in an API directory, this massively reduces the size and complexity of the index.js file, and makes it easier to add/remove/maintain APIs in general, as well as documentation.

This is ONLY Step 1.
Step 2 will involve changing the actual *endpoint* routes to match their files, and add backwards-compat. for v1.1.5 to give services time to upgrade.
